контроль редактирования на странице преподов

+++ client/src/views/Teachers.vue
@@ -1,4 +1,7 @@
 <script>
+import TeacherRow from '@/components/TeacherRow.vue';
+import _ from 'lodash';
+import Modal from '../components/modal-window.vue'
 export default {
   components: {
     Modal,
@@ -11,6 +14,7 @@ export default {
         name: "",
         patr: "",
       },
+      isEdit: false,
     };
   },
   methods: {
@@ -21,10 +25,17 @@ export default {
         this.newTeacher.name,
         this.newTeacher.patr
       );
-      this.newTeacher = {
+      this.resetTeachers();
+    },
+    onModalClose() {
+      this.showModal = false;
+      this.resetTeachers();
+    },
+    resetStudent() {
+      this.resetTeachers = {
         surname: "",
         name: "",
-        patr: "",
+        patr: ""
       };
     },
   },
@@ -34,17 +45,10 @@ export default {
 import { storeToRefs } from 'pinia';
 import {computed, onBeforeMount, ref} from 'vue';
 import {useLessonsStore} from '../stores/lessons';
-import TeacherRow from '@/components/TeacherRow.vue';
-import _ from 'lodash';
-import Modal from '../components/modal-window.vue'
 
 const lessonsStore = useLessonsStore(); 
 const {teachers} = storeToRefs(lessonsStore);
 let sortFiled = ref("name");
-
-let surname = ref("");
-let name = ref("");
-let patr = ref("");
 
 const teacherStored = computed(() =>{
     return _(teachers.value)
@@ -56,24 +60,17 @@ function toggleSort(fildeName){
     sortFiled.value = fildeName
 }
 
-// function onSurnameClick(teacher){
-//     console.log(teacher.surname)
-// }
-
-// function onNameClick(teacher){
-//     console.log(teacher.name)
-// }
-
 function onDeleteClick(teacher){
-    lessonsStore.deleteTeacher(teacher.id)
+   let permission = window.confirm(
+    `Вы действительно хотите удалить учителя ${teacher.surname} ${teacher.name} ${teacher.patr}?`
+  );
+  if (permission) {
+    lessonsStore.deleteTeacher(teacher.id);
+  }
 }
 
-// function onFormSumbit(){
-//     lessonsStore.addTeacher(surname.value, name.value, patr.value)
-// }
-
-function onUpdateClick(id, event){
-    lessonsStore.updTeacher(id,  event.name, event.surname, event.patr )
+function onUpdateClick(id, { name, surname, patr }){
+    lessonsStore.updTeacher(id,  name, surname, patr )
 }
 
 
@@ -93,7 +90,7 @@ onBeforeMount( () => {
     <modal
       :show="showModal"
       @submit="onFormSumbit()"
-      @close="showModal = false"
+      @close="onModalClose()"
     >
       <template #header>
         <h3>Добавить преподавателя</h3>
@@ -111,9 +108,20 @@ onBeforeMount( () => {
 
 <hr />
 
+    <div class="container">
     <button @click="toggleSort('name')">По имени</button>
     <button @click="toggleSort('surname')">По фамилии</button>
-    
+
+    <div class="toggle">
+      <input
+        type="checkbox"
+        id="toggle-button"
+        class="toggle-button"
+        v-model="isEdit"
+      />
+      <label for="toggle-button" class="text">Режим редактирования</label>
+    </div>
+  </div>
             <TeacherRow 
               v-for="s in teacherStored" 
               :name = "s.name"
@@ -123,11 +131,64 @@ onBeforeMount( () => {
               @name-click = "onNameClick(s)"
               @delete = "onDeleteClick(s)"
               @update = "onUpdateClick(s.id, $event)"
+              :isEdit="isEdit"
             />
     <hr> 
 
  
 </template>
-<style>
 
+<style scoped>
+.container {
+  display: flex;
+  gap: 16px;
+  padding: 8px 0;
+}
+.toggle {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+}
+.toggle-button {
+  position: relative;
+  display: inline-block;
+  width: 50px;
+  height: 25px;
+  margin: 0;
+  vertical-align: top;
+  background: #ffffff;
+  border: 1px solid #bbc1e1;
+  border-radius: 30px;
+  outline: none;
+  cursor: pointer;
+  -webkit-appearance: none;
+  -moz-appearance: none;
+  appearance: none;
+  transition: all 0.3s cubic-bezier(0.2, 0.85, 0.32, 1.2);
+}
+
+.toggle-button::after {
+  content: "";
+
+  display: inline-block;
+  position: absolute;
+  left: 3px;
+  top: 1.5px;
+
+  width: 20px;
+  height: 20px;
+  background-color: #42b983;
+  border-radius: 50%;
+
+  transform: translateX(0);
+  transition: all 0.3s cubic-bezier(0.2, 0.85, 0.32, 1.2);
+}
+
+.toggle-button:checked::after {
+  transform: translateX(calc(100% + 3px));
+  background-color: #fff;
+}
+.toggle-button:checked {
+  background-color: #42b983;
+}
 </style>
